speech to text and execution&Run  commands-online
To run the following code, Python version 3.12 must be installed on your system.
https://www.python.org/ftp/python/3.12.4/python-3.12.4-amd64.exe
https://www.python.org/ftp/python/3.12.4/python-3.12.4.exe

This code uses the following three libraries:
*speech-recognition
*sounddevice
*numpy
*soundfile

To install each of these libraries, open the Command Prompt (CMD) from the Start menu on your computer and run the following commands (it is recommended to use a VPN):

pip install SpeechRecognition
pip install sounddevice
pip install numpy

In this code, you can use the duration variable to set the recording time for capturing the speaker’s voice. After the recording, all spoken words will be recognized, and the corresponding function will be called based on the detected word.

You can also modify the functions to execute specific commands on your gadget.